separate Calibration form Calibration solver. The later still inherits from Calibration
